Move postseason logic to separate file
